feat: refactor `simple_air` test
