+++ hyperplonk/tests/simple_air.rs
@@ -2,24 +2,21 @@ use core::iter::repeat_with;
 
 use itertools::{Itertools, chain, cloned};
 use p3_air::{Air, AirBuilder, AirBuilderWithPublicValues, BaseAir, BaseAirWithPublicValues};
-use p3_challenger::{FieldChallenger, HashChallenger, SerializingChallenger32};
+use p3_air_ext::ProverInput;
 use p3_field::Field;
 use p3_field::extension::BinomialExtensionField;
-use p3_hyperplonk::{
-    ProverFolderOnExtension, ProverFolderOnExtensionPacking, ProverFolderOnPacking, ProverInput,
-    SymbolicAirBuilder, VerifierFolder, prove, verify,
-};
-use p3_keccak::Keccak256Hash;
 use p3_koala_bear::KoalaBear;
 use p3_matrix::Matrix;
 use p3_matrix::dense::RowMajorMatrix;
 use rand::distr::{Distribution, StandardUniform};
 use rand::rngs::StdRng;
 use rand::{Rng, RngCore, SeedableRng};
+use util::run;
+
+mod util;
 
 type Val = KoalaBear;
 type Challenge = BinomialExtensionField<Val, 4>;
-type Challenger = SerializingChallenger32<Val, HashChallenger<u8, Keccak256Hash, 32>>;
 
 #[derive(Clone, Copy)]
 enum MyAir {
@@ -92,38 +89,6 @@ impl MyAir {
     }
 }
 
-#[allow(clippy::multiple_bound_locations)]
-fn run<
-    #[cfg(feature = "check-constraints")] A: for<'a> Air<p3_air_ext::DebugConstraintBuilder<'a, Val>>,
-    #[cfg(not(feature = "check-constraints"))] A,
->(
-    prover_inputs: Vec<ProverInput<Val, A>>,
-) where
-    A: Clone
-        + BaseAirWithPublicValues<Val>
-        + Air<SymbolicAirBuilder<Val>>
-        + for<'t> Air<ProverFolderOnPacking<'t, Val, Challenge>>
-        + for<'t> Air<ProverFolderOnExtension<'t, Val, Challenge>>
-        + for<'t> Air<ProverFolderOnExtensionPacking<'t, Val, Challenge>>
-        + for<'t> Air<VerifierFolder<'t, Val, Challenge>>,
-{
-    let verifier_inputs = prover_inputs
-        .iter()
-        .map(|input| input.to_verifier_input())
-        .collect();
-
-    let mut prover_challenger = Challenger::from_hasher(Vec::new(), Keccak256Hash {});
-    let proof = prove(prover_inputs, &mut prover_challenger);
-
-    let mut verifier_challenger = Challenger::from_hasher(Vec::new(), Keccak256Hash {});
-    verify::<_, Challenge, _>(verifier_inputs, &proof, &mut verifier_challenger).unwrap();
-
-    assert_eq!(
-        prover_challenger.sample_algebra_element::<Challenge>(),
-        verifier_challenger.sample_algebra_element::<Challenge>(),
-    );
-}
-
 #[test]
 fn single_sum() {
     let mut rng = StdRng::from_os_rng();
@@ -131,7 +96,7 @@ fn single_sum() {
         let air = MyAir::GrandSum { width };
         let (trace, output) = air.generate_trace_rows(num_vars, &mut rng);
         let public_values = vec![output];
-        run(vec![ProverInput::new(air, public_values, trace)]);
+        run::<Val, Challenge, _>(vec![ProverInput::new(air, public_values, trace)]);
     }
 }
 
@@ -142,7 +107,7 @@ fn single_product() {
         let air = MyAir::GrandProduct { width };
         let (trace, output) = air.generate_trace_rows(num_vars, &mut rng);
         let public_values = vec![output];
-        run(vec![ProverInput::new(air, public_values, trace)]);
+        run::<Val, Challenge, _>(vec![ProverInput::new(air, public_values, trace)]);
     }
 }
 
@@ -151,18 +116,20 @@ fn multiple_mixed() {
     let mut rng = StdRng::from_os_rng();
     for _ in 0..100 {
         let n = rng.random_range(1..10);
-        run((0..n)
-            .map(|_| {
-                let num_vars = rng.random_range(0..12);
-                let width = rng.random_range(1..5);
-                let air = match rng.random_bool(0.5) {
-                    false => MyAir::GrandSum { width },
-                    true => MyAir::GrandProduct { width },
-                };
-                let (trace, output) = air.generate_trace_rows(num_vars, &mut rng);
-                let public_values = vec![output];
-                ProverInput::new(air, public_values, trace)
-            })
-            .collect());
+        run::<Val, Challenge, _>(
+            (0..n)
+                .map(|_| {
+                    let num_vars = rng.random_range(0..12);
+                    let width = rng.random_range(1..5);
+                    let air = match rng.random_bool(0.5) {
+                        false => MyAir::GrandSum { width },
+                        true => MyAir::GrandProduct { width },
+                    };
+                    let (trace, output) = air.generate_trace_rows(num_vars, &mut rng);
+                    let public_values = vec![output];
+                    ProverInput::new(air, public_values, trace)
+                })
+                .collect(),
+        );
     }
 }
